Thoughts on advanced scheduling #49
1) the update_actions is not working, you forgot to add it in the ajax
switch. Please double check all the actions to be in place, there is a
big plugin and will have a huge impact if we have any problem
2) you have increased the log checking time. Please change it back.
3) on reset action please clear also the buffer for each network.
4) please group the timeline in days and for each post display just the
time. For instance we will have a div for each day where we have posts
to be shared, and in that div we have the posts. smth like this:
http://prntscr.com/avfqca . I dont care about the design, i would like
to have the layout in place.
5) The first post when starting sharing is done after 15s. In the
advanced schedule this edge case is not treated and we get different
time here http://prntscr.com/avhdsi